Remove CodeSection.java
