+++ src/main/java/ch/njol/skript/Skript.java
@@ -1227,7 +1227,6 @@ public final class Skript extends JavaPlugin implements Listener {
 	 * @param section The section's class
 	 * @param patterns Skript patterns to match this section
 	 * @see Section
-	 * @see ch.njol.skript.lang.util.CodeSection
 	 */
 	public static <E extends Section> void registerSection(Class<E> section, String... patterns) throws IllegalArgumentException {
 		checkAcceptRegistrations();
