+++ src/main/java/ch/njol/skript/sections/SecConditional.java
@@ -22,9 +22,9 @@ import ch.njol.skript.Skript;
 import ch.njol.skript.config.SectionNode;
 import ch.njol.skript.lang.Condition;
 import ch.njol.skript.lang.Expression;
+import ch.njol.skript.lang.Section;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.skript.lang.TriggerItem;
-import ch.njol.skript.lang.util.CodeSection;
 import ch.njol.util.Kleenean;
 import org.bukkit.event.Event;
 import org.eclipse.jdt.annotation.Nullable;
@@ -33,7 +33,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 @SuppressWarnings("NotNullFieldNotInitialized")
-public class SecConditional extends CodeSection {
+public class SecConditional extends Section {
 
 	static {
 		Skript.registerSection(SecConditional.class,
